Ensure each file can be required separately
