+++ lib/concurrent-ruby-edge/concurrent/actor/behaviour/errors_on_unknown_message.rb
@@ -1,3 +1,5 @@
+require 'concurrent/actor/behaviour/abstract'
+
 module Concurrent
   module Actor
     module Behaviour
