+++ lib/concurrent-ruby-edge/concurrent/actor/default_dead_letter_handler.rb
@@ -1,3 +1,5 @@
+require 'concurrent/actor/context'
+
 module Concurrent
   module Actor
     class DefaultDeadLetterHandler < RestartingContext
