+++ lib/concurrent-ruby-edge/concurrent/actor/envelope.rb
@@ -1,3 +1,5 @@
+require 'concurrent/actor/type_check'
+
 module Concurrent
   module Actor
     class Envelope
