+++ lib/concurrent-ruby-edge/concurrent/actor/internal_delegations.rb
@@ -1,3 +1,6 @@
+require 'logger'
+require 'concurrent/actor/public_delegations'
+
 module Concurrent
   module Actor
     module InternalDelegations
