+++ lib/concurrent-ruby-edge/concurrent/actor/utils/balancer.rb
@@ -1,3 +1,5 @@
+require 'concurrent/actor/context'
+
 module Concurrent
   module Actor
     module Utils
