+++ lib/concurrent-ruby-edge/concurrent/actor/utils/pool.rb
@@ -1,4 +1,5 @@
 require 'concurrent/actor/utils/balancer'
+require 'concurrent/actor/context'
 
 module Concurrent
   module Actor
