+++ lib/concurrent-ruby-edge/concurrent/edge/old_channel_integration.rb
@@ -1,3 +1,5 @@
+require 'concurrent/promises'
+
 module Concurrent
   module Promises
     module FactoryMethods
