+++ lib/concurrent-ruby/concurrent/atomic/atomic_markable_reference.rb
@@ -1,3 +1,4 @@
+require 'concurrent/errors'
 require 'concurrent/synchronization/object'
 
 module Concurrent
