+++ lib/concurrent-ruby/concurrent/collection/lock_free_stack.rb
@@ -1,3 +1,5 @@
+require 'concurrent/synchronization/object'
+
 module Concurrent
 
   # @!macro warn.edge
