+++ lib/concurrent-ruby/concurrent/errors.rb
@@ -66,4 +66,9 @@ module Concurrent
     end
   end
 
+  # @!macro internal_implementation_note
+  class ConcurrentUpdateError < ThreadError
+    # frozen pre-allocated backtrace to speed ConcurrentUpdateError
+    CONC_UP_ERR_BACKTRACE = ['backtrace elided; set verbose to enable'].freeze
+  end
 end
