+++ lib/concurrent-ruby/concurrent/synchronization/mutex_lockable_object.rb
@@ -1,3 +1,5 @@
+require 'concurrent/synchronization/abstract_lockable_object'
+
 module Concurrent
   # noinspection RubyInstanceVariableNamingConvention
   module Synchronization
