+++ lib/concurrent-ruby/concurrent/thread_safe/util/cheap_lockable.rb
@@ -1,5 +1,6 @@
 require 'concurrent/thread_safe/util'
 require 'concurrent/thread_safe/util/volatile'
+require 'concurrent/utility/engine'
 
 module Concurrent
